added .isClosed() method to DataStore.  created H2IndexedFileRange implementation
git-svn-id: svn+ssh://svn.code.sf.net/p/jillion/code/trunk@358 d1f56cec-0f3b-4699-95c7-aedf87c10f2a

+++ test/org/jcvi/datastore/TestDefaultAceFileDataStore.java
@@ -42,7 +42,7 @@ public class TestDefaultAceFileDataStore extends TestAbstractAceParserMatchesAce
     protected List<AceContig> getContigList(File aceFile)
             throws IOException {
         DefaultAceFileDataStore dataStore= new DefaultAceFileDataStore(aceFile);
-        List<AceContig> contigs = new ArrayList<AceContig>(dataStore.size());
+        List<AceContig> contigs = new ArrayList<AceContig>((int)dataStore.size());
         for(Iterator<String> iter = dataStore.getIds(); iter.hasNext();){
             String id = iter.next();
             try {
